# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI4TJU campus guide",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/chat": {
      "post": {
        "tags": [
          "model-runtime"
        ],
        "summary": "Chat",
        "operationId": "chat_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "499": {
            "description": "Additional Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runtime/events": {
      "get": {
        "tags": [
          "model-runtime"
        ],
        "summary": "Events",
        "operationId": "events_api_runtime_events_get",
        "parameters": [
          {
            "name": "request_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "499": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Additional Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Implemented"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/requests/{request_id}/cancel": {
      "post": {
        "tags": [
          "model-runtime"
        ],
        "summary": "Cancel",
        "operationId": "cancel_api_requests__request_id__cancel_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "499": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Additional Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Implemented"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/scene/ack": {
      "post": {
        "tags": [
          "model-runtime"
        ],
        "summary": "Ack",
        "operationId": "ack_api_scene_ack_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneAck"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneAckResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "499": {
            "description": "Additional Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runtime/client-events": {
      "post": {
        "tags": [
          "model-runtime"
        ],
        "summary": "Client Events",
        "operationId": "client_events_api_runtime_client_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientEventInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeEvent"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "499": {
            "description": "Additional Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiError"
      },
...
      "CancelRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "CancelRequest"
      },
      "CancelResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "cancel_requested",
              "already_terminal"
            ],
            "title": "Status"
          },
          "local_task_stopped": {
            "type": "boolean",
            "title": "Local Task Stopped"
          },
          "upstream_stop": {
            "type": "string",
            "enum": [
              "not_started",
              "unconfirmed",
              "confirmed"
            ],
            "title": "Upstream Stop"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "status",
          "local_task_stopped",
          "upstream_stop"
        ],
        "title": "CancelResponse"
      },
      "ChatRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Message"
          },
          "mode": {
            "type": "string",
            "enum": [
              "campus_qa",
              "content_generation",
              "general_chat"
            ],
            "title": "Mode"
          },
          "campus_id": {
            "type": "string",
            "enum": [
              "weijinlu",
              "beiyangyuan"
            ],
            "title": "Campus Id"
          },
          "selected_building_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Building Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "session_id",
          "message",
          "mode",
          "campus_id"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Usage"
              },
              {
                "type": "null"
              }
            ]
          },
          "elapsed_ms": {
            "type": "number",
            "minimum": 0.0,
            "title": "Elapsed Ms"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/SceneAction"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "session_id",
          "answer",
          "sources",
          "model",
          "usage",
          "elapsed_ms",
          "actions"
        ],
        "title": "ChatResponse"
      },
      "ClientEventData": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "not_implemented",
                  "playback_failed",
                  "permission_denied",
                  "stopped"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ClientEventData"
      },
      "ClientEventInput": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "stage": {
            "type": "string",
            "enum": [
              "speech",
              "avatar"
            ],
            "title": "Stage"
          },
          "status": {
            "type": "string",
            "enum": [
              "started",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 3600000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "data": {
            "$ref": "#/components/schemas/ClientEventData"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "request_id",
          "session_id",
          "stage",
          "status",
          "data"
        ],
        "title": "ClientEventInput"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id",
          "retryable"
        ],
        "title": "ErrorDetail"
      },
...
      "EventPage": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/RuntimeEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "next_cursor": {
            "type": "integer",
            "title": "Next Cursor"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "events",
          "next_cursor",
          "truncated"
        ],
        "title": "EventPage"
      },
...
      "RuntimeEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "origin": {
            "type": "string",
            "enum": [
              "backend",
              "frontend"
            ],
            "title": "Origin"
          },
          "stage": {
            "type": "string",
            "enum": [
              "request",
              "knowledge",
              "model",
              "speech",
              "avatar",
              "scene"
            ],
            "title": "Stage"
          },
          "status": {
            "type": "string",
            "enum": [
              "started",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "data": {
            "$ref": "#/components/schemas/EventData"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "request_id",
          "seq",
          "timestamp",
          "origin",
          "stage",
          "status",
          "duration_ms",
          "data"
        ],
        "title": "RuntimeEvent"
      },
      "SceneAck": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "action_id": {
            "type": "string",
            "format": "uuid",
            "title": "Action Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "execution_failed",
                  "unsupported"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "session_id",
          "action_id",
          "status"
        ],
        "title": "SceneAck"
      },
      "SceneAckResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "action_id": {
            "type": "string",
            "format": "uuid",
            "title": "Action Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "recorded",
              "duplicate"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "action_id",
          "status"
        ],
        "title": "SceneAckResponse"
      },
      "SceneAction": {
        "properties": {
          "action_id": {
            "type": "string",
            "format": "uuid",
            "title": "Action Id"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "focus_building",
              "show_building_card"
            ],
            "title": "Type"
          },
          "parameters": {
            "$ref": "#/components/schemas/ActionParameters"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action_id",
          "request_id",
          "type",
          "parameters"
        ],
        "title": "SceneAction"
      },
...
      "Source": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "campus_id": {
            "type": "string",
            "enum": [
              "weijinlu",
              "beiyangyuan"
            ],
            "title": "Campus Id"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "retrieved_at": {
            "type": "string",
            "title": "Retrieved At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "snippet",
          "url",
          "campus_id",
          "published_at",
          "retrieved_at"
        ],
        "title": "Source"
      },
...
      "Usage": {
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Prompt Tokens"
          },
          "completion_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Completion Tokens"
          },
          "total_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Tokens"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "prompt_tokens",
          "completion_tokens",
          "total_tokens"
        ],
        "title": "Usage"
      },
...
    }
  }
}